feat(migrate): execute database rollback on rollback (#774)
closes #699
- run database rollback on `ghost update --rollback`
- forward --version to knex-migrator
  - added bridge to execute knex-migrator with the `--version` only for Ghost 2.0
