+++ lib/commands/setup.js
@@ -44,7 +44,7 @@ class SetupCommand extends Command {
         const semver = require('semver');
 
         const linux = require('../tasks/linux');
-        const migrate = require('../tasks/migrate');
+        const migrator = require('../tasks/migrator');
 
         if (argv.local) {
             argv.url = argv.url || 'http://localhost:2368/';
@@ -83,7 +83,7 @@ class SetupCommand extends Command {
 
                 // Special-case migrations
                 if (argv.stages.includes('migrate')) {
-                    tasks.push({title: 'Running database migrations', task: migrate})
+                    tasks.push({title: 'Running database migrations', task: migrator.migrate})
                 }
 
                 if (argv.stages.includes('linux-user')) {
@@ -177,7 +177,7 @@ class SetupCommand extends Command {
                 // Tack on db migration task to the end
                 tasks.push({
                     title: 'Running database migrations',
-                    task: migrate,
+                    task: migrator.migrate,
                     // CASE: We are about to install Ghost 2.0. We moved the execution of knex-migrator into Ghost.
                     enabled: () => {
                         if (semver.satisfies(instance.cliConfig.get('active-version'), '^2.0.0')) {
