+++ lib/commands/update.js
@@ -18,7 +18,7 @@ class UpdateCommand extends Command {
         const semver = require('semver');
 
         const MigrateCommand = require('./migrate');
-        const migrate = require('../tasks/migrate');
+        const migrator = require('../tasks/migrator');
 
         const instance = this.system.getInstance();
 
@@ -68,12 +68,16 @@ class UpdateCommand extends Command {
                 enabled: () => isRunning,
                 task: this.stop.bind(this)
             }, {
+                title: 'Rolling back database migrations',
+                enabled: (ctx) => ctx.rollback,
+                task: migrator.rollback
+            }, {
                 title: 'Linking latest Ghost and recording versions',
                 task: this.link
             }, {
                 title: 'Running database migrations',
                 skip: (ctx) => ctx.rollback,
-                task: migrate,
+                task: migrator.migrate,
                 // CASE: We have moved the execution of knex-migrator into Ghost 2.0.0.
                 //       If you are already on ^2 or you update from ^1 to ^2, then skip the task.
                 enabled: () => {
